+++ src/civiltekk_yugioh_scraper/v1/prod/__init__.py
@@ -1,5 +1,5 @@
-from .YGOInventoryUpload import upload_inventory_main
-from .YGOInventoryExport import export_inventory_excel, combine_ae_price
+from .ygo_inventory_upload import upload_inventory_main
+from .ygo_inventory_export import export_inventory_excel, combine_ae_price
 from .card_list_scraper import card_list_scraper
 from .bigwebscrape import bigweb_scrape
 from .yuyuteiscrape2 import yuyutei_scrape
